# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/adm/emp": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Позволяет удалить работодателя из системы. Доступ имеют только пользователи роли ADMIN",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN"
                ],
                "summary": "Удаление аккаунта работодателя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID работодателя, которого нужно удалить",
                        "name": "empID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает статус и краткую информацию ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.StatusInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Возвращает ошибку, если не удалось получить данные из запроса",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "401": {
                        "description": "Возвращает ошибку, если у пользователя нету доступа к этому функционалу.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "500": {
                        "description": "Возвращает ошибку, если на сервере произошла непредвиденная ошибка.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    }
                }
            }
        },
        "/adm/exp": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Позволяет удалить запись из системы. Доступ имеют только пользователи роли ADMIN",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN"
                ],
                "summary": "Удаление опыта",
                "parameters": [
                    {
                        "type": "string",
                        "description": "наименование записи, которую нужно удалить",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает статус и краткую информацию ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.StatusInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Возвращает ошибку, если не удалось получить данные из запроса",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "401": {
                        "description": "Возвращает ошибку, если у пользователя нету доступа к этому функционалу.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "500": {
                        "description": "Возвращает ошибку, если на сервере произошла непредвиденная ошибка.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    }
                }
            }
        },
        "/adm/status": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Позволяет удалить запись из системы. Доступ имеют только пользователи роли ADMIN",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN"
                ],
                "summary": "Удаление статуса",
                "parameters": [
                    {
                        "type": "string",
                        "description": "наименование записи, которую нужно удалить",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает статус и краткую информацию ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.StatusInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Возвращает ошибку, если не удалось получить данные из запроса",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "401": {
                        "description": "Возвращает ошибку, если у пользователя нету доступа к этому функционалу.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "500": {
                        "description": "Возвращает ошибку, если на сервере произошла непредвиденная ошибка.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    }
                }
            }
        },
        "/adm/user": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Позволяет удалить соискателя из системы. Доступ имеют только пользователи роли ADMIN",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN"
                ],
                "summary": "Удаление аккаунта соискателя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя, которого нужно удалить",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает статус и краткую информацию ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.StatusInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Возвращает ошибку, если не удалось получить данные из запроса",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "401": {
                        "description": "Возвращает ошибку, если у пользователя нету доступа к этому функционалу.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "500": {
                        "description": "Возвращает ошибку, если на сервере произошла непредвиденная ошибка.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    }
                }
            }
        },
...
        "/emp/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Позволяет получить всю основную информацию про всех работодатлей. Доступно только пользователям с ролью ADMIN",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Получить информцию про всех работодателей",
                "responses": {
                    "200": {
                        "description": "Возвращает статус 'Ok!' и массив всех данных о работодателях",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.SuccessEmployer"
                            }
                        }
                    },
                    "400": {
                        "description": "Возвращает ошибку, если не удалось получить данные из запроса (токен или передача каких-либо других данных)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "401": {
                        "description": "Возвращает ошибку, если у пользователя нету доступа к этому функционалу.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "500": {
                        "description": "Возвращает ошибку, если на сервере произошла непредвиденная ошибка.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    }
                }
            }
        },
        "/emp/auth": {
            "get": {
                "description": "Позволяет получить новый токен для работодателя, чтобы у него сохранился доступ к функционалу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Авторизовать работодателя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email работодателя",
                        "name": "Email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password работодателя",
                        "name": "Password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает статус 'Ok!', данные работодателя и новый токен",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.ResponseCreateEmployee"
                            }
                        }
                    },
                    "500": {
                        "description": "Возвращает ошибку, если на сервере произошла непредвиденная ошибка.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main_go_internal_api_Struct.InfoError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "main_go_internal_api_Struct.StatusInfo": {
            "type": "object",
            "properties": {
                "info": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "main_go_internal_api_Struct.SuccessEmployer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inn": {
                    "type": "string"
                },
                "nameOrganization": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "status": {
                    "type": "object",
                    "properties": {
                        "createdAt": {
                            "type": "string"
                        },
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}